fix: no error when loading undefined schedule name
