+++ apps/antalmanac/src/components/AppBar/ImportStudyList.tsx
@@ -22,15 +22,9 @@ import TermSelector from '../RightPane/CoursePane/SearchForm/TermSelector';
 import RightPaneStore from '../RightPane/RightPaneStore';
 import { addCustomEvent, openSnackbar } from '$actions/AppStoreActions';
 import analyticsEnum, { logAnalytics } from '$lib/analytics';
-import {
-    ZotCourseResponse,
-    addCoursesMultiple,
-    combineSOCObjects,
-    getCourseInfo,
-    queryWebsoc,
-    queryZotCourse,
-} from '$lib/helpers';
+import { ZotCourseResponse, addCoursesMultiple, queryZotCourse } from '$lib/helpers';
 import AppStore from '$stores/AppStore';
+import { getCourseInfo, queryWebsoc } from '$lib/course-helpers';
 
 const styles = {
     inputLabel: {
