+++ apps/antalmanac/src/components/RightPane/CoursePane/CourseRenderPane.tsx
@@ -17,8 +17,9 @@ import loadingGif from './SearchForm/Gifs/loading.gif';
 import darkNoNothing from './static/dark-no_results.png';
 import noNothing from './static/no_results.png';
 import AppStore from '$stores/AppStore';
-import { isDarkMode, queryWebsoc, queryWebsocMultiple } from '$lib/helpers';
+import { isDarkMode, queryWebsocMultiple } from '$lib/helpers';
 import analyticsEnum from '$lib/analytics';
+import { queryWebsoc } from '$lib/course-helpers';
 
 const styles: Styles<Theme, object> = (theme) => ({
     course: {
