+++ apps/antalmanac/tests/schedule.test.ts
@@ -1,0 +1,10 @@
+import { describe, test, expect } from 'vitest';
+import { Schedules } from '../src/stores/Schedules';
+
+describe('schedule logic', () => {
+    const scheduleStore = new Schedules();
+
+    test('no error when loading undefined schedule', () => {
+        expect(() => scheduleStore.getScheduleName(69)).not.toThrowError();
+    });
+});
